changes for post request for the rest of the report flow
