+++ rocky/reports/templates/forms/report_form_fields.html
@@ -1,7 +1,7 @@
 <input type="hidden"
        name="observed_at"
        value="{{ observed_at|date:"Y-m-d" }}">
-{% for ooi in selected_oois %}<input type="hidden" name="ooi" value={{ ooi }}>{% endfor %}
+{% for ooi in selected_oois %}<input type="hidden" name="ooi" value="{{ ooi }}">{% endfor %}
 {% for ooi_type in ooi_types_selection %}<input type="hidden" name="ooi_type" value="{{ ooi_type }}">{% endfor %}
 {% for clearance_level in clearance_levels_selection %}
     <input type="hidden" name="clearance_level" value="{{ clearance_level }}">
@@ -9,9 +9,14 @@
 {% for clearance_type in clearance_types_selection %}
     <input type="hidden" name="clearance_type" value="{{ clearance_type }}">
 {% endfor %}
-{% if plugins.required %}
-    {% for required_plugin in plugins.required %}
-        <input type="hidden" name="plugin" value={{ required_plugin.id }}>
+{% if selected_report_types %}
+    {% for report_type in selected_report_types %}
+        <input type="hidden" name="report_type" value="{{ report_type }}">
     {% endfor %}
 {% endif %}
-{% if request.GET.page %}<input type="hidden" name="page" value={{ request.GET.page }}>{% endif %}
+{% if plugins.required %}
+    {% for required_plugin in plugins.required %}
+        <input type="hidden" name="plugin" value="{{ required_plugin.id }}">
+    {% endfor %}
+{% endif %}
+{% if request.GET.page %}<input type="hidden" name="page" value="{{ request.GET.page }}">{% endif %}
