+++ rocky/reports/templates/partials/report_setup_scan.html
@@ -94,6 +94,8 @@
             {% endif %}
             <form id="generate_report" class="inline" method="post" action="{{ next }}">
                 {% csrf_token %}
+                {% include "forms/report_form_fields.html" %}
+
                 <button type="submit">
                     {% translate "Generate report" %}<span class="icon ti-chevron-right"></span>
                 </button>
