+++ rocky/reports/views/aggregate_report.py
@@ -1,16 +1,12 @@
-from collections.abc import Sequence
 from typing import Any
 
-from django.contrib import messages
 from django.http import HttpRequest, HttpResponse
 from django.shortcuts import redirect
 from django.urls import reverse
 from django.utils.http import urlencode
 from django.utils.translation import gettext_lazy as _
 
-from octopoes.models import ScanProfileType
-from octopoes.models.ooi.reports import Report as ReportOOI
-from reports.report_types.aggregate_organisation_report.report import AggregateOrganisationReport, aggregate_reports
+from reports.report_types.aggregate_organisation_report.report import AggregateOrganisationReport
 from reports.report_types.definitions import Report
 from reports.report_types.helpers import get_ooi_types_from_aggregate_report, get_report_types_from_aggregate_report
 from reports.views.base import (
@@ -21,6 +17,7 @@ from reports.views.base import (
     ReportTypeSelectionView,
     get_selection,
 )
+from reports.views.mixins import SaveAggregateReportMixin
 from reports.views.view_helpers import AggregateReportStepsMixin
 from rocky.views.ooi_view import BaseOOIListView
 
@@ -134,72 +131,6 @@ class ReportTypesSelectionAggregateReportView(
         return context
 
 
-class SaveAggregateReportMixin(ReportPluginView):
-    def save_report(self) -> ReportOOI:
-        input_oois = self.get_oois()
-
-        aggregate_report, post_processed_data, report_data, report_errors = aggregate_reports(
-            self.octopoes_api_connector,
-            input_oois,
-            self.selected_report_types,
-            self.observed_at,
-        )
-
-        # If OOI could not be found or the date is incorrect, it will be shown to the user as a message error
-        if report_errors:
-            report_types = ", ".join(set(report_errors))
-            date = self.observed_at.date()
-            error_message = _("No data could be found for %(report_types). Object(s) did not exist on %(date)s.") % {
-                "report_types": report_types,
-                "date": date,
-            }
-            messages.add_message(self.request, messages.ERROR, error_message)
-
-        observed_at = self.get_observed_at()
-
-        post_processed_data["plugins"] = self.get_plugin_data_for_saving()
-        post_processed_data["oois"] = []
-        for input_ooi in input_oois:
-            post_processed_data["oois"].append(
-                {
-                    "name": input_ooi.human_readable,
-                    "type": input_ooi.object_type,
-                    "scan_profile_level": input_ooi.scan_profile.level.value if input_ooi.scan_profile else 0,
-                    "scan_profile_type": (
-                        input_ooi.scan_profile.scan_profile_type if input_ooi.scan_profile else ScanProfileType.EMPTY
-                    ),
-                }
-            )
-
-        post_processed_data["report_types"] = []
-        for report_type in self.report_types:
-            post_processed_data["report_types"].append(
-                {
-                    "name": str(report_type.name),
-                    "description": str(report_type.description),
-                    "label_style": report_type.label_style,
-                }
-            )
-
-        # Create the report
-        report_data_raw_id = self.save_report_raw(data=post_processed_data)
-        report_ooi = self.save_report_ooi(
-            report_data_raw_id=report_data_raw_id,
-            report_type=type(aggregate_report),
-            input_oois=[ooi.primary_key for ooi in input_oois],
-            parent=None,
-            has_parent=False,
-            observed_at=observed_at,
-        )
-
-        # Save the child reports to bytes
-        for ooi, types in report_data.items():
-            for report_type, data in types.items():
-                self.save_report_raw(data=data)
-
-        return report_ooi
-
-
 class SetupScanAggregateReportView(
     SaveAggregateReportMixin, AggregateReportStepsMixin, BreadcrumbsAggregateReportView, ReportPluginView
 ):
@@ -212,11 +143,6 @@ class SetupScanAggregateReportView(
     current_step = 3
 
     def get(self, request: HttpRequest, *args: Any, **kwargs: Any) -> HttpResponse:
-        if not self.selected_report_types:
-            messages.error(self.request, _("Select at least one report type to proceed."))
-            return redirect(
-                reverse("aggregate_report_select_report_types", kwargs=self.get_kwargs()) + get_selection(self.request)
-            )
         if not self.report_has_required_plugins() or self.plugins_enabled():
             report_ooi = self.save_report()
 
@@ -230,15 +156,20 @@ class SetupScanAggregateReportView(
 
         return super().get(request, *args, **kwargs)
 
+    def post(self, request, *args, **kwargs):
+        if not self.report_type_selection_is_valid():
+            return redirect(self.get_previous())
+        return self.get(request, *args, **kwargs)
+
 
 class SaveAggregateReportView(SaveAggregateReportMixin, BreadcrumbsAggregateReportView, ReportPluginView):
     """
     Save the report and redirect to the saved report
     """
 
-    current_step = 6
-    ooi_types = get_ooi_types_from_aggregate_report(AggregateOrganisationReport)
-    report_types: Sequence[type[Report]]
+    template_name = "aggregate_report.html"
+    breadcrumbs_step = 6
+    current_step = 4
 
     def post(self, request, *args, **kwargs):
         report_ooi = self.save_report()
